+++ packages/config/src/index.ts
@@ -1,6 +1,7 @@
 import { dashboard, initDashboard } from '@graphql-portal/dashboard';
 import { prefixLogger } from '@graphql-portal/logger';
 import { ApiDef, GatewayConfig } from '@graphql-portal/types';
+import cluster from 'cluster';
 import { customAlphabet } from 'nanoid';
 import { loadApiDefs as loadApiDefsFromFs } from './api-def.config';
 import { loadConfig } from './gateway.config';
@@ -8,25 +9,27 @@ import useEnv from './use-env';
 
 const logger = prefixLogger('config');
 
-let config: {
+export type Config = {
   nodeId: string;
   gateway: GatewayConfig;
   apiDefs: ApiDef[];
   timestamp: number;
-} = {} as any;
+};
 
-export async function initConfig() {
+const config: Config = {} as any;
+
+export async function initConfig(): Promise<void> {
   config.nodeId = customAlphabet('0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz-', 11)();
   config.gateway = (await loadConfig()) as GatewayConfig;
   if (!config.gateway) {
-    throw new Error('Gateway config was not found, open config/gateway.yaml');
+    throw new Error('Gateway config was not found at path "./config/gateway.yaml"');
   }
   useEnv(config.gateway);
 }
 
-export async function loadApiDefs() {
-  if (!config.gateway) {
-    return;
+export async function loadApiDefs(): Promise<boolean> {
+  if (cluster.isWorker) {
+    return false;
   }
 
   config.apiDefs = config.apiDefs ?? [];
@@ -35,9 +38,9 @@ export async function loadApiDefs() {
     initDashboard(config.gateway);
     const loaded = await dashboard.loadApiDefs();
 
-    if (!(loaded && loaded?.apiDefs?.length)) {
+    if (!(loaded && loaded?.apiDefs)) {
       logger.info('API Definitions were not updated from Dashboard.');
-      return;
+      return false;
     }
 
     config.apiDefs = loaded.apiDefs;
@@ -47,6 +50,7 @@ export async function loadApiDefs() {
     config.timestamp = Date.now();
     useEnv(config.apiDefs);
   }
+  return true;
 }
 
 export { config };