+++ packages/gateway/src/middleware/graphql-rate-limit/rate-limit.ts
@@ -5,6 +5,7 @@ import { GraphQLError, parse, TypeInfo, ValidationContext, visit, Visitor, visit
 import CostAnalysis from 'graphql-cost-analysis/dist/costAnalysis';
 import depthLimit from 'graphql-depth-limit';
 import { apiSchema } from '../../server/router';
+import { isIntrospectionRequest, throwError } from '../utils';
 import RequestCostTool from './request-cost.tool';
 
 const logger = prefixLogger('cost-analysis');
@@ -27,16 +28,11 @@ const rateLimitMiddleware = (apiDef: ApiDef): RequestHandler => {
   logger.debug(`max cost: ${maxCost}`);
   logger.debug(`rate: ${costRate} per ${costRatePer} second(s)`);
 
-  const throwError = (error: Error, statusCode = 400): void => {
-    error.statusCode = statusCode;
-    throw error;
-  };
-
   const requestCostTool = new RequestCostTool(costRatePer);
 
   return async function rateLimitMiddleware(req: Request, res: Response, next: NextFunction): Promise<void> {
     const { body } = req;
-    if (req.method !== 'POST' || !body?.query || body?.operationName === 'IntrospectionQuery') {
+    if (isIntrospectionRequest(req)) {
       return next();
     }